split out io
