+++ tests/test_utils.py
@@ -4,6 +4,7 @@ import numpy as np
 import pytest
 
 from dolphin import utils
+from dolphin.io import load_gdal
 
 
 def test_get_dates():
@@ -33,7 +34,7 @@ def test_get_types():
 
 
 def test_get_raster_xysize(raster_100_by_200):
-    arr = utils.load_gdal(raster_100_by_200)
+    arr = load_gdal(raster_100_by_200)
     assert arr.shape == (100, 200)
     assert (200, 100) == utils.get_raster_xysize(raster_100_by_200)
